refactor(component): Extract `FlowbiteTheme` type to a file
Create a strict type for `FlowbiteTheme`.
